Add a script to compute all molecular properties
I am computing some properties of molecules for the prelims. Neocles has allowed
me to start early on this task. I have written a script which computes and
outputs all values for all molecules we are interested in.

I got overly clever with how it works but that's the fun of it. It introspects
the module molecular_properties to find all functions we are interested and then
does a topological sorting to get them in a computable order and then computes
all values.

This is a preliminary version in the future I'm going to have this write a latex
file with the table to use and a csv for plotting some values.